fix(ios): render S60v3 game icons correctly on 5320
Three S60v3 games showed broken icons on rm-409 (5320): Brothers in
Arms had a magenta backdrop, 7Days and 天地道 rendered nothing. Two
independent bugs.

1. The iOS icon decoders never applied the paired mask, so a masked
   colour-key (magenta) backdrop stayed opaque. Composite the mask onto
   the icon's alpha, with polarity chosen by mask type: a grayscale mask
   (gray2/gray256) is the alpha directly (white = opaque), while a colour
   bitmap reused as a colour-key mask (via make_standard_mask) inverts so
   white = transparent. Without the split, Final Battle's gray256 mask
   would invert and cut the icon out over magenta.

2. 7Days / 天地道 store their icon as a base64 PNG embedded in the MIF's
   SVG (<image xlink:href="data:...base64,...">), which the old lunasvg
   2.3.8 fork could not render. Merge the EKA2L1/lunasvg fork up to
   upstream v3.5.0 (native <image>/data-URI via bundled plutovg); the
   project still depends on the fork. 3.5.0 is a full rewrite without the
   fork's old <switch> stub, and these Adobe-exported icons wrap <image>
   in <switch><foreignObject .../></switch>; lunasvg skips an unknown
   element with its whole subtree, so the switch support is re-added in
   the fork (rendered as a group) rather than worked around here. The
   fork's only other downstream changes were a C++11 downgrade (moot on
   3.5.0/C++20) and a move ctor already upstreamed; icon.cpp's render
   matrix is narrowed to float for the 3.5.0 API.

Verified on 5320: all three game icons render, control icons (Final
Battle, Angry Birds, Asphalt 6, …) unregressed; Release standard
regression 11/11.

+++ src/emu/services/src/ui/icon/icon.cpp
@@ -303,8 +303,8 @@ namespace eka2l1 {
             compute_icon_size(req_w, req_h, nw, nh, fw, fh);
 
             out_rgba.assign(static_cast<std::size_t>(fw) * fh * 4, 0);
-            const double sx = (doc->width() > 0) ? (static_cast<double>(fw) / doc->width()) : 1.0;
-            const double sy = (doc->height() > 0) ? (static_cast<double>(fh) / doc->height()) : 1.0;
+            const float sx = (doc->width() > 0) ? (static_cast<float>(fw) / doc->width()) : 1.0f;
+            const float sy = (doc->height() > 0) ? (static_cast<float>(fh) / doc->height()) : 1.0f;
             lunasvg::Bitmap luna_bmp(out_rgba.data(), fw, fh, fw * 4);
             doc->render(luna_bmp, lunasvg::Matrix{ sx, 0, 0, sy, 0, 0 });
             luna_bmp.convertToRGBA();
